Add files via upload
A newly opened private university in Sri Lanka has contracted you as the Design and develop their Student Records System for Faculty of Computing. They have given you the following normal processes as their requirements for the proposed system.
To initially enroll at the university, the student provides their personal details (name, address, sex, and born year, specialization along with 5 marks. (e.g. IT,SE. DS,CS,CSN.IM) in which they
wish to enroll. A student record is created, and a student ID number is auto generated to the student. (This number is a random number generated from the system).If the student ID doesn’t contain at least four digits including the prefix as “IT”, then the system should generate a custom exception called “IDInvalidException”exception and then it required a new student ID to be created. The necessary detail also needs to be displayed in the webpage for weekday and weekend students.
Since “IDgeneration”and “showDetils”might be useful for other faculties, they will be keeping it as a separate file, so that whenever necessary, it can be use for other faculties as well. Computing Students also will use these two methods for generating the ID and the display their details.
A student can be a week end student or weekday student. Weekend student has the special attribute called “job title” and the weekday student has a special attribute called “free day”.
The system has to store the Degree Specialization details including the specialization (e.g IT,SE,DS,CSN,IM) and the GPA criteria as key value pairs.
When a student wants to enroll, the student id and the specialization will be stored as a key value pair. And this process does not require any order. ”dispalyEnrolledStudents” will display the semester enrolled student details.
A staff member can insert student details and the specialization details to the system.
Then, those details will be displayed.
All the student-enrollments should take place through the Registrar's office at the university.